parsing, validation and conversion

+++ src/main/java/com/example/sbjconvert/model/ResponseEntry.java
@@ -1,6 +1,4 @@
 package com.example.sbjconvert.model;
 
-import java.math.BigDecimal;
-
-public record ResponseEntry(String name, String transport, BigDecimal topSpeed) {
+public record ResponseEntry(String name, String transport, Double topSpeed) {
 }
